move routines to common.py
